PR feedback: move the copied packlet to build/lfx-shared (instead of temp/lfx-shared)
