+++ apps/lockfile-explorer/src/graph/test/lfxGraphLoader54.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { IJsonLfxWorkspace } from '../../../temp/lfx-shared';
+import type { IJsonLfxWorkspace } from '../../../build/lfx-shared';
 
 import * as graphTestHelpers from './graphTestHelpers';
