+++ apps/lockfile-explorer/src/graph/test/serializeToJson.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { lfxGraphSerializer, type LfxGraph } from '../../../temp/lfx-shared';
+import { lfxGraphSerializer, type LfxGraph } from '../../../build/lfx-shared';
 
 import * as lfxGraphLoader from '../lfxGraphLoader';
 import { TEST_WORKSPACE, TEST_LOCKFILE } from './testLockfile';
