+++ apps/lockfile-explorer/src/graph/test/testLockfile.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { IJsonLfxWorkspace } from '../../../temp/lfx-shared';
+import type { IJsonLfxWorkspace } from '../../../build/lfx-shared';
 
 export const TEST_WORKSPACE: IJsonLfxWorkspace = {
   workspaceRootFolder: '/test',
